+++ docs/vitepress_docs/.vitepress/components/autoDocs/CodeDocumentation.vue
@@ -11,6 +11,7 @@
         </div>
         <div class="space-y-4">
             <Card
+                class="p-2"
                 v-for="subPkg in subPackagesToDisplay"
                 :key="subPkg.module_path"
             >
@@ -23,14 +24,13 @@
         </div>
     </div>
 
-    <!-- Render class or module docs -->
-    <Card v-else-if="classData || packageData" class="mb-4">
+    <Card v-else-if="classData || packageData" class="mb-4 p-1 border-none">
         <template #default>
             <!-- Class Documentation -->
             <div v-if="classData">
                 <div class="flex items-center space-x-2 mb-2">
                     <Badge type="warning">class</Badge>
-                    <h2 class="font-mono text-lg">{{ classData.name }}</h2>
+                    <h3 class="font-mono text-xl">{{ classData.name }}</h3>
                 </div>
                 <p class="text-sm text-vp-text-2 mb-4">From: <code>{{ classData.module_path }}</code></p>
                 <Docstring class="py-3" :doc="classData.docstring"/>
@@ -46,16 +46,15 @@
                 </Accordion>
             </div>
 
-            <!-- Module Documentation -->
-            <div v-else-if="packageData">
-                <div class="flex items-center space-x-2 mb-2">
+            <div v-else>
+                <div class="flex items-center space-x-2 mb-1">
                     <Badge type="primary">module</Badge>
-                    <h2 class="font-mono text-lg">{{ packageId }}</h2>
+                    <h2 class="font-mono text-xl">{{ packageId }}</h2>
                 </div>
                 <Docstring class="py-3" :doc="packageData.module_docstring"/>
 
                 <div v-if="showAllClasses && packageData.classes.length" class="mt-4 space-y-3">
-                    <h3 class="text-base font-semibold">Classes</h3>
+                    <h2 class="text-xl font-semibold mb-4">Classes</h2>
                     <div v-for="cls in packageData.classes" :key="cls.name">
                         <CodeDocumentation
                             :class-id="`${packageId.replace(/\//g, '.')}.${cls.name}`"
@@ -65,7 +64,7 @@
                 </div>
 
                 <div v-if="showAllFunctions && packageData.functions.length" class="mt-4">
-                    <h3 class="text-base font-semibold mb-2">Functions</h3>
+                    <h2 class="text-xl font-semibold mb-4">Functions</h2>
                     <Accordion :items="functionItems">
                         <template
                             v-for="(func, i) in packageData.functions"
@@ -83,12 +82,12 @@
 
 <script lang="ts" setup>
 import {computed} from 'vue'
-import {useApiDocs} from '../composables/useApiDocs'
+import {useApiDocs} from '../../composables/useApiDocs'
 import FunctionDoc from './FunctionDoc.vue'
 import Docstring from './Docstring.vue'
-import Badge from './core/Badge.vue'
-import Card from './core/Card.vue'
-import Accordion from './core/Accordion.vue'
+import Badge from '../core/Badge.vue'
+import Card from '../core/Card.vue'
+import Accordion from '../core/Accordion.vue'
 import CodeDocumentation from './CodeDocumentation.vue'
 
 interface Props {
@@ -145,9 +144,3 @@ const subPackagesToDisplay = computed(() => {
     )
 })
 </script>
-
-<style scoped>
-.font-mono {
-    font-family: monospace;
-}
-</style>
