+++ docs/vitepress_docs/.vitepress/components/product/ProductCards.vue
@@ -1,11 +1,12 @@
 <template>
   <div class="my-8">
     <h2 class="!mb-6 !text-4xl font-bold text-center">{{ title }}</h2>
-    <div class="grid gap-6 grid-cols-1 sm:grid-cols-2 lg:grid-cols-3">
+    <div class="grid gap-6 grid-cols-1 xl:grid-cols-3 ">
       <ProductCard
         v-for="product in products"
         :key="product.name"
         :product="product"
+        class="md:px-16 md:max-w-sm w-full md:mx-auto lg:max-w-md xl:max-w-none xl:px-0 xl:mx-0"
       />
     </div>
   </div>
