+++ docs/vitepress_docs/.vitepress/theme/index.ts
@@ -11,7 +11,7 @@ import ProductCards from '../components/product/ProductCards.vue'
 import OTAIPredictionDemo from '../components/predictionDemo/OTAIPredictionDemo.vue'
 import ServicePackages from '../components/ServicePackages.vue'
 import SupportPlans from '../components/SupportPlans.vue'
-import CodeDocumentation from '../components/CodeDocumentation.vue'
+import CodeDocumentation from '../components/autoDocs/CodeDocumentation.vue'
 import Button from '../components/core/Button.vue'
 import Card from '../components/core/Card.vue'
 import Badge from '../components/core/Badge.vue'
