+++ docs/vitepress_docs/stories/Docstring.stories.ts
@@ -1,4 +1,4 @@
-import Docstring from '../.vitepress/components/Docstring.vue'
+import Docstring from '../.vitepress/components/autoDocs/Docstring.vue'
 import type { Meta, StoryObj } from '@storybook/vue3'
 
 const meta: Meta<typeof Docstring> = {
